GitHub Health Files: Code Owners
GitHub introduced [3] the "Code Owners" feature [1] back in 2017 that
allows to define individuals or teams that are responsible for code in a
repository. The feature is configured by a `codeowners` file [2] that
must be placed either in the root, `/docs` or `.github` directory of an
repository.

For @svengreb projects this file is now provided globally by this
repository for all repositories that do not require a specific or
special configuration.

[1]: https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners
[2]: https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners#codeowners-file-location
[3]: https://github.blog/2017-07-06-introducing-code-owners

GH-2