GitHub Health Files: Code Of Conduct
As part of the community health files [1], a "code of conduct" [2] has
been created to define community standards, signal a welcoming and
inclusive project, and outline procedures for handling abuse [3].
The feature is configured by a `code_of_conduct.md` file [4] that can
be placed either in the root, `/docs` or `.github` directory of an
repository.

For @svengreb projects this file is now provided globally for all
projects by this repository that do not require a specific or special
configuration.

[1]: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file
[2]: https://en.wikipedia.org/wiki/Code_of_conduct
[3]: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-code-of-conduct-to-your-project
[4]: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-code-of-conduct-to-your-project#adding-a-code-of-conduct-manually

GH-4